drm/i915/dg1: Reserve first 1MB of local memory
On DG1 A0 stepping the first 1MB of local memory must be reserved. One
reason for this is that the 0xA0000-0xB0000 range is not accessible by
the display, probably since this region is redirected to another memory
location for legacy VGA compatibility.

BSpec: 60029
Testcase: igt/kms_big_fb/linear-64bpp-rotate-0
Signed-off-by: Imre Deak <imre.deak@intel.com>