drm/i915/lmem: Do not check r->sgt.pfn for NULL
When user space does mmap, kernel would map the physical page of local memory
to virtual memory address. The r->sgt.pfn is page address allocated from
local memory and the local memory region is from 0 to LMEM size. Hence the
r->sgt.pfn is possible to be 0, but this is normal case.

Signed-off-by: Kui Wen <kui.wen@intel.com>