r515: expose logging buffers via debugfs
The LOGINIT, LOGRM, and LOGPMU buffers are circular buffers that have
printf-like logs from GSP-RM and PMU encoded in them.

LOGINIT and LOGRM are allocated by Nouveau and their DMA addresses
are passed to GSP-RM during initialization.  The buffers are required
for GSP-RM to initialize properly.

LOGPMU is also allocated by Nouveau, but its contents are updated
when Nouveau receives a NV_VGPU_MSG_EVENT_UCODE_LIBOS_PRINT RPC from
GSP-RM.  Nouveau then copies the RPC to the buffer.

The messages are encoded as an array of variable-length structures that
contain the parameters to an NV_PRINTF call.  The format string and
parameter count are stored in a special ELF image that contains only
logging strings.  This image is not currently shipped with the Nvidia
driver.

There are two methods to extract the logs.

OpenRM tries to load the logging ELF, and if present, parses the log
buffers in real time and outputs the strings to the kernel console.

Alternatively, and this is the method used by this patch, the buffers
can be exposed to user space, and a user-space tool (along with the
logging ELF image) can parse the buffer and dump the logs.

This method has the advantage that it allows the buffers to be parsed
even when the logging ELF file is not available to the user.

However, it has the disadvantage the debubfs entries need to remain
until the driver is unloaded.

The buffers are exposed via debugfs.  The debugfs entries must be
created before GSP-RM is started, to ensure that they are available
during GSP-RM initialization.

An end-user can capture the logs using the following commands:

    cp /sys/kernel/debug/nouveau/loginit loginit
    cp /sys/kernel/debug/nouveau/logrm logrm
    cp /sys/kernel/debug/nouveau/logpmu logpmu

Since LOGPMU is not needed for GSP-RM initialization, it is only created
if debugfs is available.  Otherwise, the NV_VGPU_MSG_EVENT_UCODE_LIBOS_PRINT
RPCs are ignored.

Signed-off-by: Timur Tabi <ttabi@nvidia.com>